refactor(onboarding): rename and move onboarding.constants to onboarding/contants
